# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Turiscan API - Smart Tourist Guide
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface City {
  id: number;
  name: string;
  department: string;
  country: string;
  description: string;
  imageUrl?: string | null;
}

export interface Place {
  id: number;
  cityId: number;
  cityName: string;
  name: string;
  shortDescription: string;
  history: string;
  imageUrl?: string | null;
  latitude: number;
  longitude: number;
  qrCode: string;
  category: string;
  visitHours?: string | null;
  visitDuration?: string | null;
  video360Url?: string | null;
}

export interface NotFoundError {
  error: string;
  message: string;
}

export type GetPlacesParams = {
  cityId?: number;
  search?: string;
};